Split qgscolorramp.h/.cpp out into qgscolorrampimp.h/.cpp
Avoids a bunch of unwanted includes throughout QGIS
