+++ python/core/core_auto.sip
@@ -22,6 +22,7 @@
 %Include auto_generated/qgsclipper.sip
 %Include auto_generated/qgscommandlineutils.sip
 %Include auto_generated/qgscolorramp.sip
+%Include auto_generated/qgscolorrampimpl.sip
 %Include auto_generated/qgscolorscheme.sip
 %Include auto_generated/qgscolorschemeregistry.sip
 %Include auto_generated/qgsconditionalstyle.sip
