+++ src/gui/qgscolorbrewercolorrampdialog.h
@@ -18,7 +18,7 @@
 
 #include <QDialog>
 #include "qgspanelwidget.h"
-#include "qgscolorramp.h"
+#include "qgscolorrampimpl.h"
 #include "ui_qgscolorbrewercolorrampwidgetbase.h"
 #include "qgis_gui.h"
 #include "qgis_sip.h"
