+++ src/gui/symbology/qgsgraduatedsymbolrendererwidget.cpp
@@ -30,7 +30,7 @@
 #include "qgsdatadefinedsizelegendwidget.h"
 #include "qgssymbol.h"
 #include "qgssymbollayerutils.h"
-#include "qgscolorramp.h"
+#include "qgscolorrampimpl.h"
 #include "qgscolorrampbutton.h"
 #include "qgsstyle.h"
 #include "qgsexpressioncontextutils.h"
